requirements.txt
ultralytics
opencv-python
pandas
matplotlib
numpy
scikit-learn
seaborn
python-dotenv
langchain
langchain-core
langchain-openai
langchain-neo4j
langchain-community
langchain-groq
streamlit
xmltodict